// front/src/providers/api-client.generated/model/appRightDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRoleDto } from './userRoleDto';


export interface AppRightDto { 
    id?: string;
    code: string;
    label?: string;
    roles?: Array<UserRoleDto>;
    order?: number;
}