// front/src/providers/api-client.generated/model/userRoleDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AppRightDto } from './appRightDto';


export interface UserRoleDto { 
    id?: number;
    role: string;
    rights?: Array<AppRightDto>;
    label?: string;
    enabled: boolean;
}